chore - move test file
